requirements.txt
numpy
pillow
matplotlib
opencv-python
pandas
requests
replicate
